Remove older schema package
