+++ spotbugs-exclude.xml
@@ -2,12 +2,6 @@
 <FindBugsFilter>
     <Match>
         <And>
-            <Class name="io.ballerina.stdlib.graphql.runtime.schema.Utils" />
-            <Bug pattern="BC_UNCONFIRMED_CAST" />
-        </And>
-    </Match>
-    <Match>
-        <And>
             <Class name="io.ballerina.stdlib.graphql.runtime.engine.Engine" />
             <Bug pattern="BC_UNCONFIRMED_CAST" />
         </And>
